admin  privileges

+++ src/resolvers/category.resolver.ts
@@ -19,19 +19,35 @@ export const categoryResolvers = {
         }
     },
     Mutation: {
-        createCategory: async (_:any, { category }:{
+        createCategory: async (_:any, { category, context }:{
             category:any
+            context: any
         }) => {
+            if (context && context.user && context.user.role !== "ADMIN") {
+                throw new GraphQLError("You are not authorized to create a product", {
+                    extensions: {
+                        code: "UNAUTHORIZED"
+                    }
+                });
+            }
             try{
             return await Category.create(category);
             }catch(error: any){
                 throw new GraphQLError(`Error in creating category: ${error.message}`);
             }
         },
-        updateCategory: async (_:any, { id, category }:{
+        updateCategory: async (_:any, { id, category,context }:{
             id: string;
             category: any;
+            context: any;
         }) => {
+            if (context && context.user && context.user.role !== "ADMIN") {
+                throw new GraphQLError("You are not authorized to create a product", {
+                    extensions: {
+                        code: "UNAUTHORIZED"
+                    }
+                });
+            }
             try{
             return await Category.findByIdAndUpdate(id, category, { new: true });
             }
@@ -39,9 +55,17 @@ export const categoryResolvers = {
                 throw new GraphQLError(`Error in updating category: ${error.message}`);
             }
         },
-        deleteCategory: async (_:any, { id }:{
+        deleteCategory: async (_:any, { id, context}:{
             id: string;
+            context: any;
         }) => {
+            if (context && context.user && context.user.role !== "ADMIN") {
+                throw new GraphQLError("You are not authorized to create a product", {
+                    extensions: {
+                        code: "UNAUTHORIZED"
+                    }
+                });
+            }
             try{
             return await Category.findByIdAndDelete(id);
             }catch(error: any){
